DFW.update() now correctly resets all the joystick values
There was a problem before where, when the robot first turns on, the
back left wheel turns at full power until the joystick connects. I
believe the root cause of the issue is that the array of joystick values
doesn't correctly reset the value of Joystick Left Vertical. This
revision fixes that issue hopefully.